+++ src/generator.ts
@@ -5,73 +5,87 @@ import invariant from "tiny-invariant"
 
 const TYPEOF_CHECKS = new Set(["number", "string", "boolean"])
 
-function isBuiltInType(ref: AGNodeRef): ref is BuiltinType {
-  return ref.ref === "Raw"
+type NativeTSType = "string" | "number" | "boolean" | "null"
+
+type BuiltinTS = {
+  kind: "BuiltinTS"
+  types: NativeTSType[]
 }
 
-type BuiltinType = { ref: "Raw"; name: string }
+type NodeRef = {
+  kind: "NodeRef"
+  name: string
+}
 
-// e.g. "MyNode" or "@MyUnion"
-type AGNodeRef =
-  | BuiltinType // e.g. `boolean`, or `string | boolean`
-  | { ref: "Node"; name: string }
-  | { ref: "NodeUnion"; name: string }
+// e.g. MyNode or string | boolean
+type TypeRef =
+  | BuiltinTS // e.g. boolean, or string | boolean
+  | NodeRef
 
-// e.g. "MyNode+" or "@MyUnion*"
-type AGRepeatedPattern =
-  | AGNodeRef
+// e.g. MyNode+
+type RepeatedTypePattern =
+  | TypeRef
   | {
-      ref: "List"
-      of: AGNodeRef
+      kind: "List"
+      of: TypeRef
       min: 0 | 1
     }
 
-// e.g. "MyNode?" or "@MyUnion*?"
-type AGPattern =
-  | AGRepeatedPattern
+// e.g. MyNode?
+type TypePattern =
+  | RepeatedTypePattern
   | {
-      ref: "Optional"
-      of: AGRepeatedPattern
+      kind: "Optional"
+      of: RepeatedTypePattern
     }
 
-// e.g. ['FloatLiteral', 'IntLiteral', '@StringExpr']
-type AGUnionDef = {
+type Def = UnionDef | NodeDef
+
+// e.g. Foo { bar: string } style
+type NodeDef = {
+  kind: "NodeDef"
   name: string
-  members: AGPattern[]
+  fieldsByName: LUT<Field>
+  fields: Field[]
 }
 
-type AGField = {
+// e.g. Foo = Bar | Qux
+type UnionDef = {
+  kind: "UnionDef"
   name: string
-  pattern: AGPattern
+  members: NodeRef[]
 }
 
-type AGNodeDef = {
+type Field = {
+  kind: "Field"
   name: string
-  fieldsByName: LUT<AGField>
-  fields: AGField[]
+  pattern: TypePattern
 }
 
-type AGSetting = AGSemanticDeclaration | AGSetStatement
-type AGSetStatement = { type: "set"; key: string; value: string }
-type AGSemanticDeclaration = AGSemanticPropertyDeclaration | AGSemanticMethodDeclaration
-type AGSemanticPropertyDeclaration = { type: "property"; name: string }
-type AGSemanticMethodDeclaration = { type: "method"; name: string }
-
-type AGDef = AGUnionDef | AGNodeDef
-
-type LUT<T> = Record<string, T>
-
-type AGGrammar = {
+type Grammar = {
+  kind: "Grammar"
   discriminator: string
-  externals: AGSemanticDeclaration[]
+  externals: SemanticDeclaration[]
   startNode: string
 
-  nodesByName: LUT<AGNodeDef>
-  nodes: AGNodeDef[] // Sorted list of nodes
+  nodeDefsByName: LUT<NodeDef>
+  nodeDefinitions: NodeDef[] // Sorted list of nodes
 
-  unionsByName: LUT<AGUnionDef>
-  unions: AGUnionDef[] // Sorted list of node unions
+  unionDefsByName: LUT<UnionDef>
+  unionDefinitions: UnionDef[] // Sorted list of node unions
+
+  // Convenience helper methods
+  // isBuiltinRef(ref: TypeRef): boolean
+  isUnionRef(ref: TypeRef): boolean
 }
+
+type Setting = SemanticDeclaration | SetStatement
+type SetStatement = { kind: "SetStatement"; type: "set"; key: string; value: string }
+type SemanticDeclaration = SemanticPropertyDeclaration | SemanticMethodDeclaration
+type SemanticPropertyDeclaration = { type: "property"; name: string }
+type SemanticMethodDeclaration = { type: "method"; name: string }
+
+type LUT<T> = Record<string, T>
 
 function takeWhile<T>(items: T[], predicate: (item: T) => boolean): T[] {
   const result = []
@@ -119,7 +133,7 @@ const grammar = ohm.grammar(String.raw`
       quotedIdentifier = "\"" identifier "\""
 
       // Keywords
-      keyword = semantic | property | method | set
+      keyword  = semantic | property | method | set
       semantic = "semantic" ~wordchar
       property = "property" ~wordchar
       method   = "method" ~wordchar
@@ -143,27 +157,29 @@ const grammar = ohm.grammar(String.raw`
         = semantic method identifier "(" ")"
 
       Def
-        = AGNodeDef
-        | AGUnionDef
+        = NodeDef
+        | UnionDef
 
-      AGNodeDef
-        = nodename "{" AGField* "}"
+      NodeDef
+        = nodename "{" Field* "}"
 
-      AGUnionDef
-        = "@" nodename "=" "|"? NonemptyListOf<AGNodeRef, "|">
+      UnionDef
+        = nodename "=" "|"? NonemptyListOf<NodeRef, "|">
 
-      AGField
-        = identifier "?"? ":" AGRepeatedPattern
+      Field
+        = identifier "?"? ":" RepeatedPattern
 
-      AGRepeatedPattern
-        = AGNodeRef ("+" | "*")?
+      RepeatedPattern
+        = TypeRef ("+" | "*")?
 
-      AGNodeRef
-        = BuiltinTypeUnion  -- builtin
-        | nodename          -- node
-        | "@" nodename      -- union
+      TypeRef
+        = BuiltinTSTypeUnion
+        | NodeRef
 
-      BuiltinTypeUnion
+      NodeRef
+        = nodename
+
+      BuiltinTSTypeUnion
         = NonemptyListOf<BuiltinType, "|">
 
       // e.g. \`boolean\`, or \`string | boolean\`
@@ -178,124 +194,138 @@ const grammar = ohm.grammar(String.raw`
 const semantics = grammar.createSemantics()
 
 semantics.addAttribute<
-  | AGGrammar
-  | AGSetting
-  | AGSemanticDeclaration
-  | AGNodeDef
-  | AGUnionDef
-  | AGField
-  | AGRepeatedPattern
-  | AGPattern
-  | BuiltinType
+  | Grammar
+  | Setting
+  | SemanticDeclaration
+  | NodeDef
+  | UnionDef
+  | Field
+  | RepeatedTypePattern
+  | TypePattern
+  | BuiltinTS
   | string
 >("ast", {
   nodename(_upper, _wordchar): string { return this.sourceString }, // prettier-ignore
   identifier(_letter, _wordchar): string { return this.sourceString }, // prettier-ignore
 
-  Start(settingList, defList): AGGrammar {
+  Start(settingList, defList): Grammar {
     // Settings
-    const statements = settingList.children.map((d) => d.ast as AGSetting)
+    const statements = settingList.children.map((d) => d.ast as Setting)
     const [settings, externals] = partition(
       statements,
-      (s): s is AGSetStatement => s.type === "set"
+      (s): s is SetStatement => s.type === "set"
     )
 
     const discriminator = settings.find((s) => s.key === "discriminator")?.value ?? "type"
 
     // Node definitions
-    const defs = defList.children.map((d) => d.ast as AGDef)
-    const unionsByName: LUT<AGUnionDef> = {}
-    const nodesByName: LUT<AGNodeDef> = {}
+    const defs = defList.children.map((d) => d.ast as Def)
+    const unionDefsByName: LUT<UnionDef> = {}
+    const nodeDefsByName: LUT<NodeDef> = {}
 
     for (const def of defs) {
-      if ("members" in def) {
-        unionsByName[def.name] = def
+      if (def.kind === "UnionDef") {
+        unionDefsByName[def.name] = def
       } else {
-        nodesByName[def.name] = def
+        nodeDefsByName[def.name] = def
       }
     }
 
     return {
+      kind: "Grammar",
+
       // The first-defined node in the document is the start node
-      startNode: Object.keys(nodesByName)[0]!,
+      startNode: Object.keys(nodeDefsByName)[0]!,
       discriminator,
       externals,
 
-      nodesByName,
-      nodes: Object.keys(nodesByName)
+      nodeDefsByName,
+      nodeDefinitions: Object.keys(nodeDefsByName)
         .sort()
-        .map((name) => nodesByName[name]!),
+        .map((name) => nodeDefsByName[name]!),
 
-      unionsByName,
-      unions: Object.keys(unionsByName)
+      unionDefsByName,
+      unionDefinitions: Object.keys(unionDefsByName)
         .sort()
-        .map((name) => unionsByName[name]!),
+        .map((name) => unionDefsByName[name]!),
+
+      // isBuiltinRef(ref) { return ref.kind === "BuiltinTS" },
+      isUnionRef(ref) {
+        return ref.kind === "NodeRef" && unionDefsByName[ref.name] !== undefined
+      },
     }
   },
 
-  SemanticPropertyDeclaration(_e, _p, identifier): AGSemanticPropertyDeclaration {
+  SemanticPropertyDeclaration(_e, _p, identifier): SemanticPropertyDeclaration {
     return { type: "property", name: identifier.ast as string }
   },
 
-  SemanticMethodDeclaration(_e, _m, identifier, _lp, _rp): AGSemanticMethodDeclaration {
+  SemanticMethodDeclaration(_e, _m, identifier, _lp, _rp): SemanticMethodDeclaration {
     return { type: "method", name: identifier.ast as string }
   },
 
-  SetStatement(_set, settingName, quotedName): AGSetStatement {
-    return { type: "set", key: settingName.sourceString, value: quotedName.ast as string }
+  SetStatement(_set, settingName, quotedName): SetStatement {
+    return {
+      kind: "SetStatement",
+      type: "set",
+      key: settingName.sourceString,
+      value: quotedName.ast as string,
+    }
   },
 
   quotedIdentifier(_lq, identifier, _rq): string {
     return identifier.ast as string
   },
 
-  AGNodeDef(name, _lbracket, fieldList, _rbracket): AGNodeDef {
-    const fields = fieldList.children.map((f) => f.ast as AGField)
+  NodeDef(name, _lbracket, fieldList, _rbracket): NodeDef {
+    const fields = fieldList.children.map((f) => f.ast as Field)
     const fieldsByName = index(fields, (f) => f.name)
     return {
+      kind: "NodeDef",
       name: name.ast as string,
       fieldsByName,
       fields,
     }
   },
 
-  AGUnionDef(_at, name, _eq, _pipe, memberList): AGUnionDef {
+  UnionDef(name, _eq, _pipe, memberList): UnionDef {
     return {
+      kind: "UnionDef",
       name: name.ast as string,
-      members: memberList.asIteration().children.map((m) => m.ast as AGPattern),
+      members: memberList.asIteration().children.map((m) => m.ast as NodeRef),
     }
   },
 
-  AGField(name, qmark, _colon, patternNode): AGField {
-    let pattern: AGPattern = patternNode.ast as AGRepeatedPattern
+  Field(name, qmark, _colon, patternNode): Field {
+    let pattern: TypePattern = patternNode.ast as RepeatedTypePattern
     if (qmark.children.length > 0) {
-      pattern = { ref: "Optional", of: pattern }
+      pattern = { kind: "Optional", of: pattern }
     }
-    return { name: name.ast as string, pattern }
+    return { kind: "Field", name: name.ast as string, pattern }
   },
 
-  AGRepeatedPattern(refNode, multiplier): AGRepeatedPattern {
-    let ref: AGRepeatedPattern = refNode.ast as AGNodeRef
+  RepeatedPattern(refNode, multiplier): RepeatedTypePattern {
+    let ref: RepeatedTypePattern = refNode.ast as TypeRef
     if (multiplier.children.length > 0) {
       const op = multiplier.children[0]!.sourceString
-      ref = { ref: "List", of: ref, min: op === "+" ? 1 : 0 }
+      ref = { kind: "List", of: ref, min: op === "+" ? 1 : 0 }
     }
     return ref
   },
 
-  AGNodeRef_node(nodename): AGNodeRef {
-    return { ref: "Node", name: nodename.ast as string }
+  NodeRef(nodename): NodeRef {
+    return { kind: "NodeRef", name: nodename.ast as string }
   },
 
-  AGNodeRef_union(_at, nodename): AGNodeRef {
-    return { ref: "NodeUnion", name: nodename.ast as string }
-  },
-
-  BuiltinTypeUnion(list): BuiltinType {
+  BuiltinTSTypeUnion(list): BuiltinTS {
     return {
-      ref: "Raw",
-      name: list.sourceString,
+      kind: "BuiltinTS",
+      types: list.asIteration().children.map((child) => child.ast as NativeTSType),
     }
+  },
+
+  BuiltinType(type): NativeTSType {
+    return type.sourceString as NativeTSType
   },
 })
 
@@ -309,10 +339,7 @@ semantics.addOperation<ohm.Node[]>("allRefs", {
   _iter(...children) {
     return children.flatMap((c) => (c.allRefs as () => ohm.Node[])())
   },
-  AGNodeRef_node(_nodename) {
-    return [this]
-  },
-  AGNodeRef_union(_at, _nodename) {
+  NodeRef(_nodename) {
     return [this]
   },
 })
@@ -322,7 +349,7 @@ semantics.addOperation<undefined>("check", {
     {
       const seen = new Set()
       for (const decl of externalDecls.children) {
-        const astNode = decl.ast as AGSemanticDeclaration
+        const astNode = decl.ast as SemanticDeclaration
         if (seen.has(astNode.name)) {
           throw new Error(
             decl.source.getLineAndColumnMessage() +
@@ -334,13 +361,11 @@ semantics.addOperation<undefined>("check", {
     }
 
     const validNames = new Set<string>()
-
-    const nodeDefs: AGNodeDef[] = []
-    const unionDefs: AGUnionDef[] = []
+    const defs: (NodeDef | UnionDef)[] = []
 
     // Do a pass over all defined nodes
     for (const def of defList.children) {
-      const astNode = def.ast as AGDef
+      const astNode = def.ast as Def
       if (validNames.has(astNode.name)) {
         throw new Error(
           def.source.getLineAndColumnMessage() +
@@ -349,52 +374,35 @@ semantics.addOperation<undefined>("check", {
       }
 
       validNames.add(astNode.name)
-      if ("members" in astNode) {
-        unionDefs.push(astNode)
-      } else {
-        nodeDefs.push(astNode)
-      }
+      defs.push(astNode)
     }
 
     // Do a pass over all node references
-    const nodeNames = nodeDefs.map((d) => d.name)
-    const unionNames = unionDefs.map((d) => d.name)
+    const declaredNames = defs.map((d) => d.name)
 
     const unused = new Set(validNames)
 
     // Remove the start node's name
-    unused.delete((this.ast as AGGrammar).startNode)
+    unused.delete((this.ast as Grammar).startNode)
 
     for (const ohmNode of (this.allRefs as () => ohm.Node[])()) {
-      const nodeRef = ohmNode.ast as AGNodeRef
+      const nodeRef = ohmNode.ast as NodeRef
       unused.delete(nodeRef.name)
 
-      // Check that all MyNode refs are valid
-      if (nodeRef.ref === "Node") {
-        if (!nodeNames.includes(nodeRef.name)) {
-          throw new Error(
-            ohmNode.source.getLineAndColumnMessage() + `Cannot find '${nodeRef.name}'`
-          )
-        }
-      }
-
-      // Check that all @MyUnion refs are valid
-      if (nodeRef.ref === "NodeUnion") {
-        if (!unionNames.includes(nodeRef.name)) {
-          throw new Error(
-            ohmNode.source.getLineAndColumnMessage() + `Cannot find '@${nodeRef.name}'`
-          )
-        }
+      // Check that all node refs are valid
+      if (!declaredNames.includes(nodeRef.name)) {
+        throw new Error(
+          ohmNode.source.getLineAndColumnMessage() + `Cannot find '${nodeRef.name}'`
+        )
       }
     }
 
     if (unused.size > 0) {
       const [first] = unused
       const name = first!
-      const def = defList.children.find((def) => (def.ast as AGDef).name === name)!
+      const def = defList.children.find((def) => (def.ast as Def).name === name)!
       throw new Error(
-        def.children[0]!.source.getLineAndColumnMessage() +
-          `Unused definition '${"members" in (def.ast as AGDef) ? "@" : ""}${name}'`
+        def.children[0]!.source.getLineAndColumnMessage() + `Unused definition '${name}'`
       )
     }
   },
@@ -405,80 +413,79 @@ function lowercaseFirst(text: string): string {
 }
 
 /**
- * Given a NodeRef instance, returns its formatted string, e.g. "@MyNode*"
+ * Given a NodeRef instance, returns its formatted string, e.g. "MyNode*"
  */
-function serializeRef(pat: AGPattern): string {
-  if (pat.ref === "Optional") {
+function serializeRef(pat: TypePattern): string {
+  if (pat.kind === "Optional") {
     return serializeRef(pat.of) + "?"
-  } else if (pat.ref === "List") {
+  } else if (pat.kind === "List") {
     const base = serializeRef(pat.of)
     if (pat.min > 0) {
       return base + "+"
     } else {
       return base + "*"
     }
-  } else if (pat.ref === "NodeUnion") {
-    return "@" + pat.name
-  } else if (pat.ref === "Node") {
+  } else if (pat.kind === "NodeRef") {
     return pat.name
   } else {
-    return pat.name
+    return pat.types.join(" | ")
   }
 }
 
-function getNodeRef(pat: AGPattern): AGNodeRef {
-  return pat.ref === "Optional"
+function getNodeRef(pat: TypePattern): TypeRef {
+  return pat.kind === "Optional"
     ? getNodeRef(pat.of)
-    : pat.ref === "List"
+    : pat.kind === "List"
       ? getNodeRef(pat.of)
       : pat
 }
 
-function getBareRef(pat: AGPattern): string {
-  return pat.ref === "Optional"
-    ? getBareRef(pat.of)
-    : pat.ref === "List"
-      ? getBareRef(pat.of)
-      : pat.ref === "Node"
+function getBareNodeRef(pat: TypePattern): string | undefined {
+  return pat.kind === "Optional"
+    ? getBareNodeRef(pat.of)
+    : pat.kind === "List"
+      ? getBareNodeRef(pat.of)
+      : pat.kind === "NodeRef"
         ? pat.name
-        : pat.ref === "NodeUnion"
-          ? pat.name
-          : pat.name
+        : undefined
 }
 
-function getBareRefTarget(pat: AGPattern): "Node" | "NodeUnion" | "Raw" {
-  return pat.ref === "Optional" || pat.ref === "List" ? getBareRefTarget(pat.of) : pat.ref
+function isBuiltinTypeRef(ref: TypeRef): ref is BuiltinTS {
+  return ref.kind === "BuiltinTS"
 }
 
-function getTypeScriptType(pat: AGPattern): string {
-  return pat.ref === "Optional"
+function isNodeRef(ref: TypeRef): ref is NodeRef {
+  return ref.kind === "NodeRef"
+}
+
+function getTypeScriptType(pat: TypePattern): string {
+  return pat.kind === "Optional"
     ? getTypeScriptType(pat.of) + " | null"
-    : pat.ref === "List"
+    : pat.kind === "List"
       ? getTypeScriptType(pat.of) + "[]"
-      : isBuiltInType(pat)
-        ? pat.name
+      : isBuiltinTypeRef(pat)
+        ? pat.types.join(" | ")
         : pat.name
 }
 
-function validate(grammar: AGGrammar) {
+function validate(grammar: Grammar) {
   // Keep track of which node names are referenced/used
   const referenced = new Set<string>()
 
-  for (const nodeUnion of grammar.unions) {
-    for (const ref of nodeUnion.members) {
-      const memberName = getBareRef(ref)
-      referenced.add(memberName)
+  for (const unionDef of grammar.unionDefinitions) {
+    for (const member of unionDef.members) {
+      referenced.add(member.name)
       invariant(
-        grammar.nodesByName[memberName] ??
-          (nodeUnion.name !== memberName && !!grammar.unionsByName[memberName]),
-        `Member "${memberName}" of union "${nodeUnion.name}" is not defined in the grammar`
+        grammar.nodeDefsByName[member.name] ??
+          (unionDef.name !== member.name && !!grammar.unionDefsByName[member.name]),
+        `Member "${member.name}" of union "${unionDef.name}" is not defined in the grammar`
       )
     }
   }
 
   const semanticFields = [...grammar.externals.map((ext) => ext.name)]
 
-  for (const node of grammar.nodes) {
+  for (const node of grammar.nodeDefinitions) {
     for (const field of node.fields) {
       invariant(
         field.name !== grammar.discriminator,
@@ -493,20 +500,22 @@ function validate(grammar: AGGrammar) {
         `Field '${node.name}.${field.name}' conflicts with semantic property/method`
       )
 
-      const bare = getBareRef(field.pattern)
+      const bare = getBareNodeRef(field.pattern)
+      if (bare === undefined) continue
+
       const base = getNodeRef(field.pattern)
       referenced.add(bare)
       invariant(
-        isBuiltInType(base) ||
-          !!grammar.unionsByName[bare] ||
-          !!grammar.nodesByName[bare],
+        isBuiltinTypeRef(base) ||
+          !!grammar.unionDefsByName[bare] ||
+          !!grammar.nodeDefsByName[bare],
         `Unknown node kind "${bare}" (in "${node.name}.${field.name}")`
       )
     }
   }
 
   // Check that all defined nodes are referenced
-  const unreferenced = new Set(grammar.nodes.map((n) => n.name))
+  const unreferenced = new Set(grammar.nodeDefinitions.map((n) => n.name))
   for (const name of referenced) {
     unreferenced.delete(name)
   }
@@ -522,7 +531,7 @@ function validate(grammar: AGGrammar) {
 
 function generateAssertParam(
   fieldName: string, // actualKindValue
-  fieldPat: AGPattern, // expectedNode
+  fieldPat: TypePattern, // expectedNode
   currentContext: string,
   discriminator: string
 ): string {
@@ -538,17 +547,17 @@ function generateAssertParam(
 }
 
 function generateTypeCheckCondition(
-  expected: AGPattern,
+  expected: TypePattern,
   actualValue: string,
   discriminator: string
 ): string {
   const conditions = []
 
-  if (expected.ref === "Optional") {
+  if (expected.kind === "Optional") {
     conditions.push(
       `${actualValue} === null || ${generateTypeCheckCondition(expected.of, actualValue, discriminator)}`
     )
-  } else if (expected.ref === "List") {
+  } else if (expected.kind === "List") {
     conditions.push(`Array.isArray(${actualValue})`)
     if (expected.min > 0) {
       conditions.push(`${actualValue}.length > 0`)
@@ -556,21 +565,18 @@ function generateTypeCheckCondition(
     conditions.push(
       `${actualValue}.every(item => ${generateTypeCheckCondition(expected.of, "item", discriminator)})`
     )
-  } else if (expected.ref === "Node" || expected.ref === "NodeUnion") {
+  } else if (expected.kind === "NodeRef") {
     conditions.push(`is${expected.name}(${actualValue})`)
-  } else if (isBuiltInType(expected)) {
+  } else if (isBuiltinTypeRef(expected)) {
     conditions.push(
-      `( ${expected.name
-        .replace(/`/g, "")
-        .split("|")
-        .flatMap((part) => {
-          part = part.trim()
-          if (TYPEOF_CHECKS.has(part)) {
-            return [`typeof ${actualValue} === ${JSON.stringify(part)}`]
-          } else if (part === "null") {
+      `( ${expected.types
+        .flatMap((tsType) => {
+          if (TYPEOF_CHECKS.has(tsType)) {
+            return [`typeof ${actualValue} === ${JSON.stringify(tsType)}`]
+          } else if (tsType === "null") {
             return [`${actualValue} === null`]
           } else {
-            console.warn(`Cannot emit runtime type check for ${part}`)
+            console.warn(`Cannot emit runtime type check for ${tsType}`)
             return []
           }
         })
@@ -583,7 +589,7 @@ function generateTypeCheckCondition(
 
 const onlyOnce = new WeakSet()
 
-function generateCommonSemanticHelpers(grammar: AGGrammar): string {
+function generateCommonSemanticHelpers(grammar: Grammar): string {
   if (onlyOnce.has(generateCommonSemanticHelpers)) return ""
   onlyOnce.add(generateCommonSemanticHelpers)
 
@@ -624,7 +630,7 @@ function generateCommonSemanticHelpers(grammar: AGGrammar): string {
         : never
 
     export interface ExhaustiveDispatchMap<T, C> {
-      ${grammar.nodes
+      ${grammar.nodeDefinitions
         .map((node) => `  ${node.name}(node: ${node.name}, context: C): T;`)
         .join("\n")}
     }
@@ -709,7 +715,7 @@ function generateCommonSemanticHelpers(grammar: AGGrammar): string {
   `
 }
 
-function generateMethodHelpers(grammar: AGGrammar): string {
+function generateMethodHelpers(grammar: Grammar): string {
   const methods = grammar.externals.filter((ext) => ext.type === "method")
   if (methods.length === 0) {
     return ""
@@ -752,7 +758,7 @@ function generateMethodHelpers(grammar: AGGrammar): string {
   `
 }
 
-function generatePropertyHelpers(grammar: AGGrammar): string {
+function generatePropertyHelpers(grammar: Grammar): string {
   const props = grammar.externals.filter((ext) => ext.type === "property")
   if (props.length === 0) {
     return ""
@@ -795,7 +801,7 @@ function generatePropertyHelpers(grammar: AGGrammar): string {
   `
 }
 
-function parseGrammarFromPath(path: string): AGGrammar {
+function parseGrammarFromPath(path: string): Grammar {
   const src = fs.readFileSync(path, "utf-8")
   return parseGrammarFromString(src)
 }
@@ -808,7 +814,7 @@ function index<T>(arr: T[], keyFn: (item: T) => string): LUT<T> {
   return result
 }
 
-export function parseGrammarFromString(text: string): AGGrammar {
+export function parseGrammarFromString(text: string): Grammar {
   const parsed = grammar.match(text)
   if (parsed.message) {
     throw new Error(parsed.message)
@@ -816,10 +822,10 @@ export function parseGrammarFromString(text: string): AGGrammar {
 
   const tree = semantics(parsed)
   ;(tree.check as () => void)() // Will throw in case of errors
-  return tree.ast as AGGrammar
+  return tree.ast as Grammar
 }
 
-function generateCode(grammar: AGGrammar): string {
+function generateCode(grammar: Grammar): string {
   // Will throw in case of errors
   validate(grammar)
 
@@ -908,16 +914,11 @@ function generateCode(grammar: AGGrammar): string {
     `,
   ]
 
-  for (const union of grammar.unions) {
-    const [subNodes, subUnions] = partition(
-      union.members,
-      (ref) => getBareRefTarget(ref) === "Node"
-    )
+  for (const union of grammar.unionDefinitions) {
+    const [subUnions, subNodes] = partition(union.members, (m) => grammar.isUnionRef(m))
     const conditions = subNodes
-      .map(
-        (ref) => `value.${grammar.discriminator} === ${JSON.stringify(getBareRef(ref))}`
-      )
-      .concat(subUnions.map((ref) => `is${getBareRef(ref)}(value)`))
+      .map((ref) => `value.${grammar.discriminator} === ${JSON.stringify(ref.name)}`)
+      .concat(subUnions.map((ref) => `is${ref.name}(value)`))
     output.push(`
       export function is${union.name}(value: unknown): value is ${union.name} {
         return isNode(value) && (${conditions.join(" || ")})
@@ -925,10 +926,10 @@ function generateCode(grammar: AGGrammar): string {
     `)
   }
 
-  for (const union of grammar.unions) {
+  for (const union of grammar.unionDefinitions) {
     output.push(`
       export type ${union.name} =
-        ${union.members.map((member) => getBareRef(member)).join(" | ")};
+        ${union.members.map((member) => member.name).join(" | ")};
     `)
   }
 
@@ -936,16 +937,16 @@ function generateCode(grammar: AGGrammar): string {
     ${generateMethodHelpers(grammar)}
     ${generatePropertyHelpers(grammar)}
 
-    export type Node = ${grammar.nodes.map((node) => node.name).join(" | ")}
+    export type Node = ${grammar.nodeDefinitions.map((node) => node.name).join(" | ")}
 
     export type ChildrenOf<N extends Node> = {
-      ${grammar.nodes
+      ${grammar.nodeDefinitions
         .map((node) => {
           const nodeType = node.name
           const childTypes = new Set(
             node.fields
               .map((f) => getNodeRef(f.pattern))
-              .filter((ref) => !isBuiltInType(ref))
+              .filter(isNodeRef)
               .map((ref) => ref.name)
           )
           return `${JSON.stringify(nodeType)}: ${[...childTypes].join(" | ") || "never"},`
@@ -985,7 +986,7 @@ function generateCode(grammar: AGGrammar): string {
     }
   `)
 
-  for (const node of grammar.nodes) {
+  for (const node of grammar.nodeDefinitions) {
     output.push(`
       export interface ${node.name} extends Semantics {
           ${grammar.discriminator}: ${JSON.stringify(node.name)}
@@ -1008,13 +1009,13 @@ function generateCode(grammar: AGGrammar): string {
   }
 
   output.push("")
-  for (const node of grammar.nodes) {
+  for (const node of grammar.nodeDefinitions) {
     const optionals = new Set(
       takeWhile(
         node.fields.slice().reverse(),
         (field) =>
-          field.pattern.ref === "Optional" ||
-          (field.pattern.ref === "List" && field.pattern.min === 0)
+          field.pattern.kind === "Optional" ||
+          (field.pattern.kind === "List" && field.pattern.min === 0)
       ).map((field) => field.name)
     )
 
@@ -1029,7 +1030,7 @@ function generateCode(grammar: AGGrammar): string {
           const key = field.name
           const type = getTypeScriptType(field.pattern)
           return optionals.has(field.name)
-            ? `${key}: ${type} = ${field.pattern.ref === "Optional" ? "null" : "[]"}`
+            ? `${key}: ${type} = ${field.pattern.kind === "Optional" ? "null" : "[]"}`
             : `${key}: ${type}`
         }),
         "range: Range = [0, 0]",
@@ -1052,9 +1053,9 @@ function generateCode(grammar: AGGrammar): string {
       switch (node.${grammar.discriminator}) {
   `)
 
-  for (const node of grammar.nodes) {
+  for (const node of grammar.nodeDefinitions) {
     const fields = node.fields.filter(
-      (field) => !isBuiltInType(getNodeRef(field.pattern))
+      (field) => !isBuiltinTypeRef(getNodeRef(field.pattern))
     )
 
     if (fields.length === 0) {
@@ -1064,9 +1065,8 @@ function generateCode(grammar: AGGrammar): string {
     output.push("")
     output.push(`case ${JSON.stringify(node.name)}:`)
     for (const field of fields) {
-      switch (field.pattern.ref) {
-        case "Node":
-        case "NodeUnion":
+      switch (field.pattern.kind) {
+        case "NodeRef":
           output.push(`  yield node.${field.name} as ChildrenOf<N>`)
           break